Added exchange and routing_key headers to SubscribeHeaders to better support AMQP brokers (such as RabbitMQ).
Fixes issue# 10

Added an optional parameter to the connect method in Stomp class for passing in a Socket.  This makes it possible to pass in custom socket implementations, such as the TLSSocket form as3crypo (see http://code.google.com/p/as3crypto).

Fixes issue#7

Previously removed all references to any classes in the Flex framework.  This makes it possible to use the Stomp lib with Flash projects

Fixes #9